# backend/openapi.yaml
components:
  schemas:
    AnalyticsRead:
      properties:
        analytics_type:
          title: Analytics Type
          type: string
        category:
          title: Category
          type: string
        id:
          title: Id
          type: integer
        metric_name:
          title: Metric Name
          type: string
        metric_value:
          title: Metric Value
          type: number
        period:
          default: daily
          title: Period
          type: string
        recorded_at:
          format: date-time
          title: Recorded At
          type: string
        user_id:
          title: User Id
          type: integer
      required:
      - analytics_type
      - category
      - metric_name
      - metric_value
      - id
      - user_id
      - recorded_at
      title: AnalyticsRead
      type: object
    AssignmentCreate:
      properties:
        attachment_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Attachment Url
        course_id:
          title: Course Id
          type: integer
        description:
          title: Description
          type: string
        due_date:
          format: date-time
          title: Due Date
          type: string
        instructions:
          anyOf:
          - type: string
          - type: 'null'
          title: Instructions
        lesson_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Lesson Id
        max_score:
          default: 100
          title: Max Score
          type: integer
        title:
          title: Title
          type: string
      required:
      - title
      - description
      - course_id
      - due_date
      title: AssignmentCreate
      type: object
    AssignmentRead:
      properties:
        attachment_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Attachment Url
        course_id:
          title: Course Id
          type: integer
        created_at:
          format: date-time
          title: Created At
          type: string
        description:
          title: Description
          type: string
        due_date:
          format: date-time
          title: Due Date
          type: string
        id:
          title: Id
          type: integer
        instructions:
          anyOf:
          - type: string
          - type: 'null'
          title: Instructions
        is_active:
          title: Is Active
          type: boolean
        lesson_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Lesson Id
        max_score:
          default: 100
          title: Max Score
          type: integer
        teacher_id:
          title: Teacher Id
          type: integer
        title:
          title: Title
          type: string
      required:
      - title
      - description
      - course_id
      - due_date
      - id
      - teacher_id
      - is_active
      - created_at
      title: AssignmentRead
      type: object
...
    GradeCreate:
      properties:
        assignment_id:
          title: Assignment Id
          type: integer
        feedback:
          anyOf:
          - type: string
          - type: 'null'
          title: Feedback
        score:
          title: Score
          type: integer
        student_id:
          title: Student Id
          type: integer
      required:
      - assignment_id
      - student_id
      - score
      title: GradeCreate
      type: object
    GradeRead:
      properties:
        assignment_id:
          title: Assignment Id
          type: integer
        feedback:
          anyOf:
          - type: string
          - type: 'null'
          title: Feedback
        graded_at:
          format: date-time
          title: Graded At
          type: string
        graded_by:
          title: Graded By
          type: integer
        id:
          title: Id
          type: integer
        score:
          title: Score
          type: integer
        student_id:
          title: Student Id
          type: integer
      required:
      - assignment_id
      - student_id
      - score
      - id
      - graded_by
      - graded_at
      title: GradeRead
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ParentDashboardRead:
      properties:
        child_id:
          title: Child Id
          type: integer
        generated_at:
          format: date-time
          title: Generated At
          type: string
        id:
          title: Id
          type: integer
        parent_id:
          title: Parent Id
          type: integer
        report_data:
          title: Report Data
          type: string
      required:
      - id
      - parent_id
      - child_id
      - report_data
      - generated_at
      title: ParentDashboardRead
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: /api/auth/login
      type: oauth2
info:
  title: Gyanvruksh API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/analytics/parent/{child_id}:
    get:
      operationId: get_parent_dashboard_api_analytics_parent__child_id__get
      parameters:
      - in: path
        name: child_id
        required: true
        schema:
          title: Child Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ParentDashboardRead'
                title: Response Get Parent Dashboard Api Analytics Parent  Child Id  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Get Parent Dashboard
      tags:
      - analytics
  /api/analytics/record:
    post:
      operationId: record_analytics_api_analytics_record_post
      requestBody:
        content:
          application/json:
            schema:
              title: Analytics Data
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Record Analytics
      tags:
      - analytics
  /api/analytics/user:
    get:
      operationId: get_user_analytics_api_analytics_user_get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AnalyticsRead'
                title: Response Get User Analytics Api Analytics User Get
                type: array
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: Get User Analytics
      tags:
      - analytics
  /api/assignments/:
    get:
      operationId: list_assignments_api_assignments__get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AssignmentRead'
                title: Response List Assignments Api Assignments  Get
                type: array
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: List Assignments
      tags:
      - assignments
    post:
      operationId: create_assignment_api_assignments__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentCreate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Create Assignment
      tags:
      - assignments
  /api/assignments/{assignment_id}/grade:
    post:
      operationId: grade_assignment_api_assignments__assignment_id__grade_post
      parameters:
      - in: path
        name: assignment_id
        required: true
        schema:
          title: Assignment Id
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeCreate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GradeRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Grade Assignment
      tags:
      - assignments
  /api/assignments/{assignment_id}/grades:
    get:
      operationId: get_grades_api_assignments__assignment_id__grades_get
      parameters:
      - in: path
        name: assignment_id
        required: true
        schema:
          title: Assignment Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/GradeRead'
                title: Response Get Grades Api Assignments  Assignment Id  Grades
                  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Get Grades
      tags:
      - assignments
...
  /api/teacher/grading/assignment:
    post:
      description: Grade a student assignment
      operationId: grade_assignment_api_teacher_grading_assignment_post
      parameters:
      - in: query
        name: student_id
        required: true
        schema:
          title: Student Id
          type: integer
      - in: query
        name: course_id
        required: true
        schema:
          title: Course Id
          type: integer
      - in: query
        name: assignment_id
        required: true
        schema:
          title: Assignment Id
          type: integer
      - in: query
        name: grade
        required: true
        schema:
          title: Grade
          type: number
      - in: query
        name: feedback
        required: true
        schema:
          title: Feedback
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Grade Assignment
      tags:
      - teacher
...